+++ NEWS.md
@@ -17,6 +17,7 @@
 * Add integration test coverage for SUDOC call-number browse type filtering (TestRail C627509)
 * Add integration test coverage for LC call-number browse with empty config returning all types (TestRail C627500)
 * Add integration test coverage for NLM call-number browse with empty config returning all types (TestRail C627501)
+* Add integration test coverage for Other scheme call-number browse with empty config returning all types (TestRail C627502)
 
 ### Dependencies
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`